+++ Editor/Processors/OriginalState.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using Anatawa12.AvatarOptimizer.ErrorReporting;
 using nadena.dev.ndmf;
 using UnityEngine;