+++ Editor/Processors/SkinnedMeshes/MergeSkinnedMeshProcessor.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Anatawa12.AvatarOptimizer.ErrorReporting;
 using nadena.dev.ndmf;
 using UnityEditor;
 using UnityEditor.Animations;
@@ -92,12 +91,10 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
             {
                 // collect (skinned) mesh renderers who doesn't have normal
                 // to show the list on the error reporting
-                BuildReport.LogFatal("MergeSkinnedMesh:error:mix-normal-existence")
-                    ?.WithContext((
-                        from meshInfo2 in meshInfos
-                        where meshInfo2.Vertices.Count != 0 && !meshInfo2.HasNormals
-                        select (object)meshInfo2.SourceRenderer
-                    ).ToArray());
+                BuildLog.LogError("MergeSkinnedMesh:error:mix-normal-existence",
+                    from meshInfo2 in meshInfos
+                    where meshInfo2.Vertices.Count != 0 && !meshInfo2.HasNormals
+                    select meshInfo2.SourceRenderer);
             }
 
             Profiler.EndSample();
@@ -208,7 +205,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
 #endif
 
             foreach (var weightMismatchBlendShape in weightMismatchBlendShapes)
-                BuildReport.LogWarning("MergeSkinnedMesh:warning:blendShapeWeightMismatch", weightMismatchBlendShape);
+                BuildLog.LogWarning("MergeSkinnedMesh:warning:blendShapeWeightMismatch", weightMismatchBlendShape);
 
             if (updateBounds && newBoundMin != Vector3.positiveInfinity && newBoundMax != Vector3.negativeInfinity)
             {
@@ -237,8 +234,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
                 var toDestroy = renderer.GetComponent<RemoveZeroSizedPolygon>();
                 if (toDestroy)
                 {
-                    BuildReport.LogWarning("MergeSkinnedMesh:warning:removeZeroSizedPolygonOnSources")
-                        ?.WithContext(toDestroy);
+                    BuildLog.LogWarning("MergeSkinnedMesh:warning:removeZeroSizedPolygonOnSources", toDestroy);
                     Object.DestroyImmediate(toDestroy);
                 }
                 Object.DestroyImmediate(renderer);
@@ -285,15 +281,14 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
 
         private void ActivenessAnimationWarning(Renderer renderer, BuildContext context, HashSet<Transform> parents)
         {
-            ErrorLog log = null;
+            var sources = new List<object>();
 
             // Warn if the source mesh can be hidden differently than merged by animation.
             {
                 if (context.GetAnimationComponent(renderer).TryGetFloat("m_Enabled", out var p))
                 {
-                    log = log ?? BuildReport.LogWarning("MergeSkinnedMesh:warning:animation-mesh-hide")
-                        ?.WithContext(renderer);
-                    log?.WithContext(p.Sources);
+                    sources.Add(renderer);
+                    sources.Add(p.SourcesEnum);
                 }
             }
             foreach (var transform in renderer.transform.ParentEnumerable(context.AvatarRootTransform, includeMe: true))
@@ -301,12 +296,14 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
                 if (parents.Contains(transform)) break;
                 if (context.GetAnimationComponent(transform.gameObject).TryGetFloat("m_IsActive", out var p))
                 {
-                    log = log ?? BuildReport.LogWarning("MergeSkinnedMesh:warning:animation-mesh-hide")
-                        ?.WithContext(renderer);
-                    log?.WithContext(transform.gameObject);
-                    log?.WithContext(p.Sources);
+                    sources.Add(renderer);
+                    sources.Add(transform.gameObject);
+                    sources.Add(p.SourcesEnum);
                 }
             }
+
+            if (sources.Count != 0)
+                BuildLog.LogWarning("MergeSkinnedMesh:warning:animation-mesh-hide", sources);
         }
 
         private (int[][] mapping, List<(MeshTopology topology, Material material)> materials)