+++ Editor/Processors/TraceAndOptimize/ComponentDependencyCollector.cs
@@ -1,11 +1,13 @@
+using System;
+using System.Collections.Generic;
 using Anatawa12.AvatarOptimizer.APIInternal;
-using Anatawa12.AvatarOptimizer.ErrorReporting;
 using Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes;
 using JetBrains.Annotations;
 using nadena.dev.ndmf;
 using UnityEditor;
 using UnityEngine;
 using Debug = System.Diagnostics.Debug;
+using Object = UnityEngine.Object;
 
 namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
 {
@@ -30,11 +32,12 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
         public void CollectAllUsages()
         {
             var collector = new Collector(this, _componentInfos);
+            var unknownComponents = new Dictionary<Type, List<Object>>();
             // second iteration: process parsers
             foreach (var componentInfo in _componentInfos.AllInformation)
             {
                 var component = componentInfo.Component;
-                BuildReport.ReportingObject(component, () =>
+                using (ErrorReport.WithContextObject(component))
                 {
                     // component requires GameObject.
                     collector.Init(componentInfo);
@@ -44,14 +47,18 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
                     }
                     else
                     {
-                        BuildReport.LogWarning("TraceAndOptimize:warn:unknown-type", component.GetType().Name);
+                        if (!unknownComponents.TryGetValue(component.GetType(), out var list))
+                            unknownComponents.Add(component.GetType(), list = new List<Object>());
+                        list.Add(component);
 
                         FallbackDependenciesParser(component, collector);
                     }
 
                     collector.FinalizeForComponent();
-                });
+                }
             }
+            foreach (var (type, objects) in unknownComponents)
+                BuildLog.LogWarning("TraceAndOptimize:warn:unknown-type", type, objects);
         }
 
         private static void FallbackDependenciesParser(Component component, API.ComponentDependencyCollector collector)