+++ Editor/Processors/UnusedBonesByReferencesToolEarlyProcessor.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using Anatawa12.AvatarOptimizer.ErrorReporting;
 using nadena.dev.ndmf;
 using UnityEngine;
 
@@ -17,12 +16,12 @@ namespace Anatawa12.AvatarOptimizer.Processors
             var configuration = context.AvatarRootObject.GetComponent<UnusedBonesByReferencesTool>();
             if (!configuration) return;
 
-            BuildReport.ReportingObject(configuration, () =>
+            using (ErrorReport.WithContextObject(configuration))
             {
                 UnusedBonesByReferences.Make(BoneReference.Make(configuration.transform,
                         configuration.detectExtraChild), configuration.preserveEndBone)
                     .SetEditorOnlyToBones();
-            });
+            }
         }
 
         #region UnusedBonesByReferencesTool